maint(resources): use actions/setup-node to upgrade npm
We don't have permissions to use `npm install --global npm` to get a
specific version, but this action is supported by GH so should be fine.

It remains to be seen if there will be other compatibility issues due to
being a later version of Node than what we are using in v19.

Build-bot: skip
Test-bot: skip