refactor(web): renames PendingTokenization as TokenizationPath, adds doc-comments
Now that the SearchSpace rework is stabilizing, I've finally landed on a better name for the type.  PendingTokenization _directly_ correlates to `SearchPath` entries once processed, so the name `TokenizationPath` will reflect this nicely.  New doc-comments are also added documenting the relationship of `TokenizationPath` and `TokenizationSubset` to `SearchPath` and `SearchCluster`.

Admittedly, it may be wise to additionally rename `TokenizationSubset` to `TokenizationCluster` (and rename the source file) to further mirror the relationship of these types... but that can always be done in a follow-up.

Build-bot: skip build:web
Test-bot: skip