fix(web): improve determination of default path 🍒
Browser extensions can dynamically insert scripts into a web page, so
it's possible that the last script is not what we expected, which
can cause the path to the script to be wrong. This change fixes the
determination of the default path by using the `document.currentScript`
property. It also replaces the use of the deprecated `substr` function
in those places.

Cherry-pick-of: #15159
Fixes: #15158
Test-bot: skip