chore(windows): upgrade sentry-native to 0.12.1
Note: this links to #15182, but investigation is still ongoing. Given
the work is largely done, it's probably worth finishing off and running
the sentry-client-test-exception tests once the symbolication issue is
addresed, and staying on the latest version.

sentry.pas has been partially updated to match sentry.h changes. The
changes that need to be made can be seen in
https://github.com/getsentry/sentry-native/compare/0.6.0...0.12.1. There
is a marker `//---UPDATED TO HERE FOR 0.12.1---` in sentry.pas to show
where I stopped.